REFACTOR: 가격 타입 변경(int -> BigDecimal) (#242)
* fix(book): 가격 타입 변경(int -> BigDecimal)

* fix(book): 가격 타입 변경(int -> BigDecimal)

* fix(book): 가격 타입 변경(int -> BigDecimal)

* feat:
-테스트 코드 반영 때문에 이전 branch의 추가된 service와 Repository 반영

* feat:
-테스트 코드 issue 때문에 임시로 이전 branch의 추가된 코드 선반영

---------

+++ src/main/java/com/nhnacademy/illuwa/d_book/book/document/BookDocument.java
@@ -10,6 +10,8 @@ import org.springframework.data.elasticsearch.annotations.Field;
 import org.springframework.data.elasticsearch.annotations.FieldType;
 import org.springframework.data.elasticsearch.annotations.Setting;
 
+import java.math.BigDecimal;
+
 @Document(indexName = "books-v1")
 @Setting(settingPath = "elasticsearch/nori-settings.json")
 @Getter
@@ -36,7 +38,7 @@ public class BookDocument {
     private String isbn;
 
     @Field(type = FieldType.Integer)
-    private int salePrice;
+    private BigDecimal salePrice;
 
     @Field(type = FieldType.Text)
     private String thumbnailUrl;
